+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fix a bug preventing DB upgrade when online update is disabled
 - Various bug fixes and improvements
 - Extend run_pvault.sh to work in Linux enviroments
+- Add ingredient stock managment
 
 ### Version 3.7
 - Delete history log when formula is deleted